requirements.txt
dataclasses==0.6
Django==2.2.7
django-cors-headers==3.1.1
djangorestframework==3.10.3
djongo==1.2.36
dnspython==1.16.0
pymongo==3.9.0
pytz==2019.3
sqlparse==0.2.4
